Rename FormatStore to formatStore
