--- src/Sulu/Bundle/MediaBundle/Resources/js/stores/FormatStore/index.js
@@ -1,4 +1,0 @@
-// @flow
-import FormatStore from './FormatStore';
-
-export default FormatStore;
