+++ src/Sulu/Bundle/MediaBundle/Resources/js/stores/formatStore/formatStore.js
