+++ src/Sulu/Bundle/MediaBundle/Resources/js/stores/formatStore/index.js
@@ -1,0 +1,4 @@
+// @flow
+import formatStore from './formatStore';
+
+export default formatStore;
